asas

+++ app/Http/Controllers/AdminUsersController.php
@@ -159,7 +159,14 @@ class AdminUsersController extends Controller
         //
         $user= User::findOrFail($id);
 
-        unlink(public_path().$user->photo->file);
+        if(!$user->photo == null)
+        {
+
+            unlink(public_path().$user->photo->file);
+
+
+        }
+
 
         $user->delete();
